Add Rest support for aliases
Even though we don't support aliases yet, we are providing some code which is used by spring-elasticsearch project for the transport client.

Because we want to provide as well the Rest implementation there, we need to support creating aliases with the Rest client. See https://github.com/dadoonet/spring-elasticsearch/issues/113

And we should test it :)

Relates to #18.

Closes #26.