No need to delete templates before updating them
If templates are already existing, you can just send the update template again to overwrite the template.
No need to remove them then update.

That creates a short period of time where a template can be removed and a create index request happen in the same time.
That will produce unexpected results.

As a consequence, the `force` parameter is not applied anymore to templates, component templates and index templates. So they are always updated.

Closes #167.